Implement getById

+++ src/main/java/com/maradamark09/typingspeedtest/word/WordServiceImpl.java
@@ -4,7 +4,6 @@ import com.maradamark09.typingspeedtest.difficulty.DifficultyNotFoundException;
 import com.maradamark09.typingspeedtest.difficulty.DifficultyRepository;
 import com.maradamark09.typingspeedtest.exception.ResourceAlreadyExistsException;
 import com.maradamark09.typingspeedtest.exception.ResourceNotFoundException;
-import com.maradamark09.typingspeedtest.util.PaginationUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -13,7 +12,7 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class WordServiceImpl implements WordService{
+public class WordServiceImpl implements WordService {
 
     private final WordRepository wordRepository;
 
@@ -21,7 +20,7 @@ public class WordServiceImpl implements WordService{
 
     @Override
     public List<String> getAllByDifficulty(String difficulty) throws ResourceNotFoundException {
-        if(!difficultyRepository.existsByValue(difficulty.toLowerCase()))
+        if (!difficultyRepository.existsByValue(difficulty.toLowerCase()))
             throw new DifficultyNotFoundException();
         return wordRepository.findAllByDifficulty(difficulty);
     }
@@ -29,34 +28,42 @@ public class WordServiceImpl implements WordService{
     @Override
     public List<String> getRandomWordsByDifficulty(String difficulty, Integer amount) throws ResourceNotFoundException {
 
-        if(!difficultyRepository.existsByValue(difficulty.toLowerCase()))
+        if (!difficultyRepository.existsByValue(difficulty.toLowerCase()))
             throw new DifficultyNotFoundException();
 
         return wordRepository.findRandomWordsByDifficulty(difficulty, Pageable.ofSize(amount));
     }
 
     @Override
-    public Word save(WordRequest wordRequest) throws ResourceNotFoundException, WordLengthGreaterThanDifficultyException, ResourceAlreadyExistsException {
+    public Word getById(Long id) throws ResourceNotFoundException {
+        return wordRepository.findById(id)
+                .orElseThrow(() -> new WordNotFoundException(id));
+    }
+
+    @Override
+    public Word save(WordRequest wordRequest)
+            throws ResourceNotFoundException, WordLengthGreaterThanDifficultyException, ResourceAlreadyExistsException {
         var difficulty = difficultyRepository.findById(wordRequest.difficulty_id())
                 .orElseThrow(() -> new DifficultyNotFoundException(wordRequest.difficulty_id()));
 
-        if(wordRequest.value().length() > difficulty.getMaxWordLength())
+        if (wordRequest.value().length() > difficulty.getMaxWordLength())
             throw new WordLengthGreaterThanDifficultyException(wordRequest.value(), difficulty);
 
-        if(wordRepository.existsByValue(wordRequest.value()))
+        if (wordRepository.existsByValue(wordRequest.value()))
             throw new WordAlreadyExistsException(wordRequest.value());
 
-        var wordToSave =
-                Word.builder()
-                        .value(wordRequest.value())
-                        .difficulty(difficulty)
-                        .build();
+        var wordToSave = Word.builder()
+                .value(wordRequest.value())
+                .difficulty(difficulty)
+                .build();
         return wordRepository.save(wordToSave);
     }
 
     @Override
     public void deleteById(Long id) throws ResourceNotFoundException {
-        wordRepository.findById(id).orElseThrow(() -> new WordNotFoundException(id));
+        if (!wordRepository.existsById(id)) {
+            throw new WordNotFoundException(id);
+        }
         wordRepository.deleteById(id);
     }
